Add files via upload
In this project I learned the basics of eventListner, callback function, switch-case and setTimeout. Used click and keydown event to play the drums.
This may seem to be not so much from a pro POV, but as a beginner its still a step forward and some confidence to carry forward.